Major re-working to ensure correct ordering
Fixes #6. This is a large change to the design. It now builds two
arrays: one of tweets, and then one of oembed objects. It can then sort
the oembed object array into correct date order and render it.

The previous method was much more concise but made no guarantees about
correct ordering, due to the indeterminate nature of asynchronous
recursion.

Note: this commit still has a lot of debugging console.log statements
in as I'd like to test it more before removing them.